HOME-162 - Decouple Agent base from Agent type
